Add function name support
